Refactor file structure of project
