+++ src/components/App.js
@@ -1,8 +1,10 @@
 import React, { Component } from "react";
 import { Jumbotron } from "react-bootstrap";
 import { Switch, Route, BrowserRouter } from "react-router-dom";
+
 import Home from "./Home";
 import User from "./User";
+
 class App extends Component {
   render() {
     return (
