+++ src/index.js
@@ -1,10 +1,13 @@
-import "bootstrap/dist/css/bootstrap.css";
-import "bootstrap/dist/css/bootstrap-theme.css";
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import registerServiceWorker from "./registerServiceWorker";
+
+import "bootstrap/dist/css/bootstrap.css";
+import "bootstrap/dist/css/bootstrap-theme.css";
+
+import App from "./components/App";
+import registerServiceWorker from "./services/registerServiceWorker";
+
+import "./assets/styles.css";
 
 ReactDOM.render(<App />, document.getElementById("root"));
 registerServiceWorker();
